seperate reader
